requirements/local.txt
-r base.txt


django-debug-toolbar==3.5.0
django-extensions==3.1.5